Weekly plan (#1)
* moves next auth migrations to standard migrations

* getMonday util function

* consent banner

* partial redux store

* Macro summaries

* removes logs

* adds zod dependency

* GET /api/plan

* GET /api/plan/[planId]/slots

* slot related api routes

* updates plan slice

* Slots DAL

* formats project

* PlanGrid base layout

* slot acitons

* recipe tag and recipe tags many to many table

* implements GET /api/recipe/[recipeId]

* implements GET /api/recipe route only

* Ingredient schemas

* typo in trg_recompute_recipe

* getRecipe specs

* formats project

* getOrCreatePlan specs

* comments

* ci workflow

* getPlanSlots specs

* updates /api/plan/[planId]/slots routes

* deleteSlot specs

* searchRecipe initial implementation

* fixes ranking and sort bugs with searchRecipe

* updates migrator eslint config

* updates control eslint config

* getInfiniteRecipes array query param handling

* validates response with recipeSearchResultsSchema in getInfiniteRecipes

* sets initial selected day to today

* chips

* recipe search api integration

* removes unused var

* translations for RecipeListEmptyState

* adds tag filters

* togglable chips

* fixes bug with debounced query

* permissions checks in create slot

* refactors components and adds FilledCell loading state

* slot upsert

* pagination controls

* linting issues

+++ packages/storage/src/migrations/1755993148328-CreateMealSlot.ts
@@ -10,7 +10,7 @@ export class CreateMealSlot1755993148328 implements MigrationInterface {
           day        int            not null,
           meal       meal_type_enum not null,
           recipe_id  uuid           not null references recipe (id) on delete restrict on update cascade,
-          servings   numeric(5, 2)  not null default 1,
+          servings   double precision  not null default 1,
           created_at timestamptz    NOT NULL DEFAULT now(),
           updated_at timestamptz    NOT NULL DEFAULT now(),
           constraint meal_day_range check (day >= 0 and day <= 6),
